Added support for different regions!
The register command now supports a second argument that specifies
the region of the player to register.
Also, there is a new region command that specifies the default region
used by the register command in case the second argument is omitted.

The pubg api access code and many commands had to be changed
to make all of this this work. Plus I commented many unit tests out
for the moment because this change made obvious that some kind
of "db access layer" is overdue, because stubbing all knex.js functions
for every single unit tests is way to much work and complicated