Added settings
The settings can be found in the new config folder and can be used to
tweak the default region and cache timings.
The optional auth.json file for easier development is now supposed to
reside inside the config folder.
The dockerignore file was changed to reflect these changes.
In future updates the settings file could be used to specify admins,
languages, roles etc. for the bot.